ajuste de base de datos

+++ src/Services/Mentoring/Mentoria.Services.Mentoring.Application/ProgramMentoring/ProgramActivitiesSolutions/Common/ProgramActivitySolutionResponse.cs
@@ -6,6 +6,6 @@ namespace Mentoria.Services.Mentoring.Application.ProgramMentoring.ProgramActivi
         Guid IdProgramActivity,
         Guid IdUser,
         Guid IdFileSolution,
-        decimal Calification
+        double Calification
     );
 }
